refactor: 팔로우 페이지 탭에 추가
