+++ app/src/Profile/index.tsx
@@ -6,7 +6,7 @@ import { BeginAtProvider } from '@/Profile/components/BeginAtProvider';
 import { UserProfile } from '@/Profile/components/UserProfile';
 import { UserProfileProvider } from '@/Profile/components/UserProfileProvider';
 
-import { ProfileTabList } from './ProfileTabList';
+import { ProfileTabList } from './components/ProfileTabList';
 
 const ProfileLayout = () => {
   return (
